web/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.0.0
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
tensorflow>=2.13.0
scikit-image>=0.19.0
joblib>=1.2.0